feat: Added Admin portal

+++ src/api/chat/delete-chat.js
@@ -2,6 +2,7 @@ import express from 'express';
 import jwtStrategy from '../../strategy/auth/jwtauth';
 import db from '../../../models';
 import crypto from 'crypto';
+import { authenticateDataForUser } from '../../authenticateDataForUser';
 
 //chatName is the name of the chat
 //isGroupChat tells whether it is to create a group or direct message chat
@@ -10,6 +11,13 @@ import crypto from 'crypto';
 let controller = async (req, res, next)=>{
     try{
         let {chatId} = req.params;
+        await authenticateDataForUser({userId: req.user.id, chatId});
+        let chat =  await db.Chat.findOne({
+            where: {
+                id: chatId
+            }
+        });
+        if(!chat) throw new RequestError("Invalid ChatId", 400);
         console.log("Reached to delete chat: ", chatId);
         await db.ChatUser.destroy({
             where: {
@@ -17,18 +25,29 @@ let controller = async (req, res, next)=>{
             },
             // force: true
         })
-        await db.Message.destroy({
-            where: {
-                chatId
-            },
-            // force: true
-        });
-        await db.Chat.destroy({
-            where: {
-                id: chatId
-            },
-            // force: true
-        });
+        if(!chat.isGroupchat){
+            await db.Message.destroy({
+                where: {
+                    chatId
+                },
+                // force: true
+            });
+            await db.Chat.destroy({
+                where: {
+                    id: chatId
+                },
+                // force: true
+            });
+            await db.Friend_Request.destroy({
+                where: {
+                    [db.Sequelize.Op.or]: [
+                        { from: chat.chatName.split('_')[0], to: chat.chatName.split('_')[1] },
+                        { from: chat.chatName.split('_')[1], to: chat.chatName.split('_')[0] }
+                    ]
+                },
+                // force: true
+            });
+        }
 
         res.status(200).json({
             success: true,
